+++ lib/DAV/Auth/Backend/AbstractBasic.php
@@ -2,8 +2,11 @@
 
 namespace Sabre\DAV\Auth\Backend;
 
-use Sabre\DAV;
-use Sabre\HTTP;
+use
+    Sabre\DAV,
+    Sabre\HTTP,
+    Sabre\HTTP\RequestInterface,
+    Sabre\HTTP\ResponseInterface;
 
 /**
  * HTTP Basic authentication backend class
@@ -20,11 +23,21 @@ use Sabre\HTTP;
 abstract class AbstractBasic implements BackendInterface {
 
     /**
-     * This variable holds the currently logged in username.
+     * Authentication Realm.
      *
-     * @var string|null
+     * The realm is often displayed by browser clients when showing the
+     * authentication dialog.
+     *
+     * @var string
      */
-    protected $currentUser;
+    protected $realm = 'sabre/dav';
+
+    /**
+     * This is the prefix that will be used to generate principal urls.
+     *
+     * @var string
+     */
+    protected $principalPrefix = 'principals/';
 
     /**
      * Validates a username and password
@@ -39,46 +52,93 @@ abstract class AbstractBasic implements BackendInterface {
     abstract protected function validateUserPass($username, $password);
 
     /**
-     * Returns information about the currently logged in username.
+     * Sets the authentication realm for this backend.
      *
-     * If nobody is currently logged in, this method should return null.
-     *
-     * @return string|null
+     * @param string $realm
+     * @return void
      */
-    function getCurrentUser() {
-        return $this->currentUser;
-    }
+    function setRealm($realm) {
 
+        $this->realm = $realm;
+
+    }
 
     /**
-     * Authenticates the user based on the current request.
+     * When this method is called, the backend must check if authentication was
+     * successful.
      *
-     * If authentication is successful, true must be returned.
-     * If authentication fails, an exception must be thrown.
+     * The returned value must be one of the following
      *
-     * @param DAV\Server $server
-     * @param string $realm
-     * @throws DAV\Exception\NotAuthenticated
-     * @return bool
+     * [true, "principals/username"]
+     * [false, "reason for failure"]
+     *
+     * If authentication was successful, it's expected that the authentication
+     * backend returns a so-called principal url.
+     *
+     * Examples of a principal url:
+     *
+     * principals/admin
+     * principals/user1
+     * principals/users/joe
+     * principals/uid/123457
+     *
+     * If you don't use WebDAV ACL (RFC3744) we recommend that you simply
+     * return a string such as:
+     *
+     * principals/users/[username]
+     *
+     * @param RequestInterface $request
+     * @param ResponseInterface $response
+     * @return array
      */
-    function authenticate(DAV\Server $server, $realm) {
+    function check(RequestInterface $request, ResponseInterface $response) {
 
-        $auth = new HTTP\Auth\Basic($realm, $server->httpRequest, $server->httpResponse);
-        $userpass = $auth->getCredentials($server->httpRequest);
+        $auth = new HTTP\Auth\Basic(
+            $this->realm,
+            $request,
+            $response
+        );
+
+        $userpass = $auth->getCredentials($request);
         if (!$userpass) {
-            $auth->requireLogin();
-            throw new DAV\Exception\NotAuthenticated('No basic authentication headers were found');
+            return [false, "No 'Authorization: Basic' header found. Either the client didn't send one, or the server is mis-configured"];
         }
-
-        // Authenticates the user
         if (!$this->validateUserPass($userpass[0],$userpass[1])) {
-            $auth->requireLogin();
-            throw new DAV\Exception\NotAuthenticated('Username or password does not match');
+            return [false, "Username or password was incorrect"];
         }
-        $this->currentUser = $userpass[0];
-        return true;
+        return [true, $this->principalPrefix . $userpass[0]];
+
     }
 
+    /**
+     * This method is called when a user could not be authenticated, and
+     * authentication was required for the current request.
+     *
+     * This gives you the oppurtunity to set authentication headers. The 401
+     * status code will already be set.
+     *
+     * In this case of Basic Auth, this would for example mean that the
+     * following header needs to be set:
+     *
+     * $response->addHeader('WWW-Authenticate', 'Basic realm=SabreDAV');
+     *
+     * Keep in mind that in the case of multiple authentication backends, other
+     * WWW-Authenticate headers may already have been set, and you'll want to
+     * append your own WWW-Authenticate header instead of overwriting the
+     * existing one.
+     *
+     * @return void
+     */
+    function requireAuth(RequestInterface $request, ResponseInterface $response) {
+
+        $auth = new HTTP\Auth\Basic(
+            $this->realm,
+            $request,
+            $response
+        );
+        $auth->requireLogin();
+
+    }
 
 }