+++ lib/DAVACL/Plugin.php
@@ -87,15 +87,6 @@ class Plugin extends DAV\ServerPlugin {
     public $hideNodesFromListings = false;
 
     /**
-     * This string is prepended to the username of the currently logged in
-     * user. This allows the plugin to determine the principal path based on
-     * the username.
-     *
-     * @var string
-     */
-    public $defaultUsernamePath = 'principals';
-
-    /**
      * This list of properties are the properties a client can search on using
      * the {DAV:}principal-property-search report.
      *
@@ -242,10 +233,7 @@ class Plugin extends DAV\ServerPlugin {
         if (is_null($authPlugin)) return null;
         /** @var $authPlugin Sabre\DAV\Auth\Plugin */
 
-        $userName = $authPlugin->getCurrentUser();
-        if (!$userName) return null;
-
-        return $this->defaultUsernamePath . '/' .  $userName;
+        return $authPlugin->getCurrentPrincipal();
 
     }