+++ tests/Sabre/DAV/Auth/Backend/ApacheTest.php
@@ -14,50 +14,59 @@ class ApacheTest extends \PHPUnit_Framework_TestCase {
 
     }
 
-    /**
-     * @expectedException Sabre\DAV\Exception
-     */
     function testNoHeader() {
 
-        $server = new DAV\Server();
+        $request = new HTTP\Request();
+        $response = new HTTP\Response();
         $backend = new Apache();
-        $backend->authenticate($server,'Realm');
+
+        $this->assertFalse(
+            $backend->check($request, $response)[0]
+        );
 
     }
 
     function testRemoteUser() {
 
-        $backend = new Apache();
-
-        $server = new DAV\Server();
         $request = HTTP\Sapi::createFromServerArray([
             'REMOTE_USER' => 'username',
         ]);
-        $server->httpRequest = $request;
+        $response = new HTTP\Response();
+        $backend = new Apache();
 
-        $this->assertTrue($backend->authenticate($server, 'Realm'));
-
-        $userInfo = 'username';
-
-        $this->assertEquals($userInfo, $backend->getCurrentUser());
+        $this->assertEquals(
+            [true, 'principals/username'],
+            $backend->check($request, $response)
+        );
 
     }
 
     function testRedirectRemoteUser() {
 
-        $backend = new Apache();
-
-        $server = new DAV\Server();
         $request = HTTP\Sapi::createFromServerArray([
             'REDIRECT_REMOTE_USER' => 'username',
         ]);
-        $server->httpRequest = $request;
+        $response = new HTTP\Response();
+        $backend = new Apache();
 
-        $this->assertTrue($backend->authenticate($server, 'Realm'));
+        $this->assertEquals(
+            [true, 'principals/username'],
+            $backend->check($request, $response)
+        );
 
-        $userInfo = 'username';
+    }
 
-        $this->assertEquals($userInfo, $backend->getCurrentUser());
+    function testRequireAuth() {
+
+        $request = new HTTP\Request();
+        $response = new HTTP\Response();
+
+        $backend = new Apache();
+        $backend->requireAuth($request, $response);
+
+        $this->assertNull(
+            $response->getHeader('WWW-Authenticate')
+        );
 
     }
 }