+++ tests/Sabre/DAV/Auth/PluginTest.php
@@ -16,6 +16,7 @@ class PluginTest extends \PHPUnit_Framework_TestCase {
         $this->assertTrue($plugin instanceof Plugin);
         $fakeServer->addPlugin($plugin);
         $this->assertEquals($plugin, $fakeServer->getPlugin('auth'));
+        $this->assertInternalType('array', $plugin->getPluginInfo());
 
     }
 
@@ -25,7 +26,7 @@ class PluginTest extends \PHPUnit_Framework_TestCase {
     function testAuthenticate() {
 
         $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
-        $plugin = new Plugin(new Backend\Mock(),'realm');
+        $plugin = new Plugin(new Backend\Mock());
         $fakeServer->addPlugin($plugin);
         $this->assertTrue(
             $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()])
@@ -40,54 +41,89 @@ class PluginTest extends \PHPUnit_Framework_TestCase {
     function testAuthenticateFail() {
 
         $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
-        $plugin = new Plugin(new Backend\Mock(),'failme');
+        $backend = new Backend\Mock();
+        $backend->fail = true;
+
+        $plugin = new Plugin($backend);
         $fakeServer->addPlugin($plugin);
         $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
 
     }
 
-    function testReportPassThrough() {
+    /**
+     * @depends testAuthenticate
+     */
+    function testMultipleBackend() {
 
-        $fakeServer = new DAV\Server(new DAV\SimpleCollection('bla'));
-        $plugin = new Plugin(new Backend\Mock(),'realm');
+        $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
+        $backend1 = new Backend\Mock();
+        $backend2 = new Backend\Mock();
+        $backend2->fail = true;
+
+        $plugin = new Plugin();
+        $plugin->addBackend($backend1);
+        $plugin->addBackend($backend2);
+
         $fakeServer->addPlugin($plugin);
+        $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
 
-        $request = HTTP\Sapi::createFromServerArray(array(
-            'REQUEST_METHOD' => 'REPORT',
-            'HTTP_CONTENT_TYPE' => 'application/xml',
-            'REQUEST_URI' => '/',
-        ));
-        $request->setBody('<?xml version="1.0"?><s:somereport xmlns:s="http://www.rooftopsolutions.nl/NS/example" />');
-
-        $fakeServer->httpRequest = $request;
-        $fakeServer->sapi = new HTTP\SapiMock();
-        $fakeServer->httpResponse = new HTTP\ResponseMock();
-        $fakeServer->exec();
-
-        $this->assertEquals(415, $fakeServer->httpResponse->status);
+        $this->assertEquals('principals/admin', $plugin->getCurrentPrincipal());
 
     }
 
     /**
      * @depends testInit
+     * @expectedException Sabre\DAV\Exception
      */
-    function testGetCurrentUserPrincipal() {
+    function testNoAuthBackend() {
 
         $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
-        $plugin = new Plugin(new Backend\Mock(),'realm');
+
+        $plugin = new Plugin();
+        $fakeServer->addPlugin($plugin);
+        $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
+
+    }
+    /**
+     * @depends testInit
+     * @expectedException Sabre\DAV\Exception
+     */
+    function testInvalidCheckResponse() {
+
+        $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
+        $backend = new Backend\Mock();
+        $backend->invalidCheckResponse = true;
+
+        $plugin = new Plugin($backend);
+        $fakeServer->addPlugin($plugin);
+        $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
+
+    }
+
+    /**
+     * @depends testAuthenticate
+     */
+    function testGetCurrentPrincipal() {
+
+        $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
+        $plugin = new Plugin(new Backend\Mock());
+        $fakeServer->addPlugin($plugin);
+        $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
+        $this->assertEquals('principals/admin', $plugin->getCurrentPrincipal());
+
+    }
+
+    /**
+     * @depends testAuthenticate
+     */
+    function testGetCurrentUser() {
+
+        $fakeServer = new DAV\Server( new DAV\SimpleCollection('bla'));
+        $plugin = new Plugin(new Backend\Mock());
         $fakeServer->addPlugin($plugin);
         $fakeServer->emit('beforeMethod', [new HTTP\Request(), new HTTP\Response()]);
         $this->assertEquals('admin', $plugin->getCurrentUser());
 
-    }
-
-    /**
-     * @depends testInit
-     */
-    function testPlugin() {
-        $myRealmName = 'some_realm';
-        $plugin = new Plugin(new Backend\Mock(),$myRealmName);
-        $this->assertEquals($myRealmName, $plugin->getRealm());
     }
 
 }